requirements.txt
x10-python-trading-starknet==0.0.6
python-dotenv==1.0.1
numpy==1.26.4
pytest-asyncio==0.23.7
